+++ test/Splunk.OpenTelemetry.AutoInstrumentation.Tests/EffectiveResourceConfigReaderTests.cs
@@ -1,0 +1,63 @@
+// <copyright file="EffectiveResourceConfigReaderTests.cs" company="Splunk Inc.">
+// Copyright Splunk Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using OpenTelemetry.Resources;
+using Splunk.OpenTelemetry.AutoInstrumentation.EffectiveConfig;
+
+namespace Splunk.OpenTelemetry.AutoInstrumentation.Tests;
+
+public class EffectiveResourceConfigReaderTests
+{
+    [Fact]
+    public void ReadServiceNameFromResource_ReturnsServiceNameAttribute()
+    {
+        var resource = ResourceBuilder.CreateEmpty()
+            .AddAttributes([new KeyValuePair<string, object>("service.name", "configured-service")])
+            .Build();
+
+        Assert.Equal("configured-service", EffectiveResourceConfigReader.ReadServiceName(resource));
+    }
+
+    [Fact]
+    public void ReadServiceNameFromResource_ReturnsNull_WhenServiceNameMissing()
+    {
+        var resource = ResourceBuilder.CreateEmpty()
+            .AddAttributes([new KeyValuePair<string, object>("service.version", "1.0.0")])
+            .Build();
+
+        Assert.Null(EffectiveResourceConfigReader.ReadServiceName(resource));
+    }
+
+    [Fact]
+    public void ReadServiceNameFromResource_ReturnsNull_WhenServiceNameEmpty()
+    {
+        var resource = ResourceBuilder.CreateEmpty()
+            .AddAttributes([new KeyValuePair<string, object>("service.name", string.Empty)])
+            .Build();
+
+        Assert.Null(EffectiveResourceConfigReader.ReadServiceName(resource));
+    }
+
+    [Fact]
+    public void ReadServiceNameFromResource_ReturnsNull_WhenServiceNameHasWrongType()
+    {
+        var resource = ResourceBuilder.CreateEmpty()
+            .AddAttributes([new KeyValuePair<string, object>("service.name", 123L)])
+            .Build();
+
+        Assert.Null(EffectiveResourceConfigReader.ReadServiceName(resource));
+    }
+}
